chore: Moved to derive message 6
